% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_render.R
\name{tt_render}
\alias{tt_render}
\title{Render a typst_table to Typst markup}
\usage{
tt_render(table)
}
\arguments{
\item{table}{A \code{typst_table} object.}
}
\value{
A character string containing Typst table markup.
}
\description{
Generates Typst code from a \code{typst_table} object. This is called automatically
by \code{print()} and \code{knit_print()}, but can also be called directly.
}
\examples{
code <- tt(mtcars[1:3, 1:3]) |> tt_render()
cat(code)

}
